+++ www/routes/businessLogic/render.js
@@ -114,6 +114,7 @@ router.post('/list', checkIfLoggedIn, function (req, res, next) {
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Search/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Search-Channel/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Search-Playlist/'));
+    promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Most-Popular/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/userspec-Others/'));
     Promise.all(promise_array).then(values => {
         directory['twitter-Tweet'] = values[0];
@@ -127,7 +128,8 @@ router.post('/list', checkIfLoggedIn, function (req, res, next) {
         directory['youtube-Search'] = values[8];
         directory['youtube-Search-Channel'] = values[9];
         directory['youtube-Search-Playlist'] = values[10];
-        directory['userspec-Others'] = values[11];
+        directory['youtube-Most-Popular'] = values[11];
+        directory['userspec-Others'] = values[12];
         res.send(directory);
     }).catch(err => {
         res.send({ERROR: err});
@@ -148,6 +150,7 @@ router.post('/list-all', checkIfLoggedIn, function (req, res, next) {
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Search/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Search-Channel/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Search-Playlist/'));
+    promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Most-Popular/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/userspec-Others/'));
     var graphqlLength = promise_array.length;
 
@@ -185,7 +188,8 @@ router.post('/list-all', checkIfLoggedIn, function (req, res, next) {
         directory['GraphQL']['youtube-Search'] = values[8];
         directory['GraphQL']['youtube-Search-Channel'] = values[9];
         directory['GraphQL']['youtube-Search-Playlist'] = values[10];
-        directory['GraphQL']['userspec-Others'] = values[11];
+        directory['GraphQL']['youtube-Most-Popular'] = values[11];
+        directory['GraphQL']['userspec-Others'] = values[12];
         for (i = 0; i < order.length; i++) {
             var parent = order[i]['parent'];
             var child = order[i]['child'];